Split tests out to separate file.
